Post-merge fix: GHOST_ContextMTL macro clashes with newly renamed GHOST_ContextMTL class
The iOS branch was so far using a GHOST_ContextMTL macro to switch
between the iOS Metal graphics context (GHOST_ContextIOS) and the macOS
Metal context (GHOST_ContextCGL). GHOST_ContextCGL was recently
renamed to GHOST_ContextMTL in main (following removal of OpenGL code)
and thus now clashes with this macro. Rename the macro to
GHOST_ContextMetal to fix this issue.