iOS: Add iPad Simulator Support
This PR brings support for the Xcode iPad iOS Simulator. As the iOS
Simulator Framework, SDK and ABI are different, this patch effectively
adds a new separate `ios-simulator` platform, and requires a separate
set of library dependencies to be built. This required the following
changes to the build system:

- Add a new `ios-simulator` `APPLE_TARGET_DEVICE`, with the required
  SDK / compile flag changes
- Add a new `APPLE_TARGET_IOS` CMake helper variable which is `TRUE` for
  both iOS and iOS Simulator builds, used for general iOS checks
- Add a new `make deps ios-simulator` command to invoke the dependency
  builder, which will be built in `../build_ios-simulator/arm64/` and
  installed in `/lib/ios-simulator_arm64/`

The dependency builder also required the following iOS library patches
modifications:
- FFmpeg: Add a new `ios-simulator` OS target, which mostly take the
  same setting as its `ios` counterpart with a compilation flag
  modification to use the right ABI, patch was also slightly
  cleaned-up
- x265: Use the `APPLE_OS_MINVERSION_CFLAG` CMake variable to use the
  right compilation flag, removing a TODO

Since separate libraries need to be built in
`/lib/ios-simulator_arm64/`, this directory was added to the
`.gitignore` to keep a clean git working tree. Since this platform is
only meant to be used by developers, I didn't see any need to make it a
precompiled library submodule like other platforms, but this could
eventually be reconsidered, let me know what you think.

See the Pull Request for details on how to build and run Blender on the
Simulator

Pull Request: https://projects.blender.org/blender/blender/pulls/143389