iOS: Inward Edge Swipe Gestures
Supported gestures:
* Inwards left edge.
* Inwards right edge.

This calls an operator (screen.edge_swipe) which is supposed to handle
all edge swipes. At the moment it only works for the viewport to bring
the Tools and Properties Region in when in Focus Mode.

I could not get outwards edge swipe to work. It requires a regular pan
gesture, and it fights with other gestures (including the edge swipe).

Co-authored-by: Jonas Holzman <jonas@holzman.fr>
Pull Request: https://projects.blender.org/blender/blender/pulls/144849